修复视频预览画面，不能能过人脸查场景图和人脸搜索

+++ UI/videoplayer.cpp
@@ -104,8 +104,8 @@ VideoPlayer::VideoPlayer(QWidget *parent):
         QDialogButtonBox::ButtonRole returnRole = QDialogButtonBox::ButtonRole(setWaringAreaDialog.exec());
         if(returnRole == QDialog::Accepted){
             QVector<QPolygonF> polygons = paintArea->polygons();
-            qreal xPer = videoWidth() / paintArea->width();
-            qreal yPer = videoHeidht() / paintArea->height();
+            qreal xPer = videoWidth() / (qreal)paintArea->width();
+            qreal yPer = videoHeidht() / (qreal)paintArea->height();
             bool isABDoor = radioBtn->isChecked() ? true : false;
             AreaType::type ty = isABDoor ? AreaType::ABDOOR : AreaType::FORBIDDENZONE;
             QVector<QPair<int,QPolygonF>> areas;
